// src/lib/api/generated/swagger-model/internalHandlerBilibiliJob.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Series API
 * Backend API for users and movies with MongoDB.
 * OpenAPI spec version: 1.0
 */
import type { InternalHandlerUploadItem } from './internalHandlerUploadItem';

export interface InternalHandlerBilibiliJob {
  createdAt?: string;
  error?: string;
  finishedAt?: string;
  id?: string;
  items?: InternalHandlerUploadItem[];
  message?: string;
  progress?: number;
  sourceUrl?: string;
  stage?: string;
  startedAt?: string;
  status?: string;
  total?: number;
}